cs3505_media_timeline/SpotifyData/requirements.txt
spotipy==2.16.1
Pillow==8.0.1
pathlib2==2.3.5
